requirements.txt
distribute==0.6.24
Django==1.4.3
djangorestframework==2.1.16
easy-thumbnails==1.2
Pillow==1.7.8
pip-tools==0.2.1
simplejson==3.1.0
South==0.7.6
